move item search to FindQtItem.cpp
