+++ lib/src/Scene/Qt/QtScene.cpp
@@ -5,6 +5,7 @@
  ****/
 
 #include "QtScene.h"
+#include "FindQtItem.h"
 
 #include <Scene/Qt/QtItem.h>
 #include <Scene/Qt/QtItemTools.h>
@@ -17,170 +18,13 @@
 #include <QQuickItem>
 #include <QQuickWindow>
 
-namespace {
-
-QQuickWindow* GetQQuickWindowWithName(const std::string& name)
-{
-    QString qtName = QString::fromStdString(name);
-    QQuickWindow* foundWindow = nullptr;
-
-    auto windows = QGuiApplication::topLevelWindows();
-    for (const auto& window : windows) {
-        QQuickWindow* qquickWindow = qobject_cast<QQuickWindow*>(window);
-        if (qquickWindow && (spix::qt::GetObjectName(qquickWindow) == qtName)) {
-            foundWindow = qquickWindow;
-            break;
-        }
-    }
-
-    return foundWindow;
-}
-
-QObject* MatchesSelector(QObject* item, const spix::path::Selector& selector)
-{
-    return std::visit(
-        [item](const auto& specific_selector) -> QObject* {
-            using SelectorType = std::decay_t<decltype(specific_selector)>;
-
-            if constexpr (std::is_same_v<SelectorType, spix::path::NameSelector>) {
-                const auto& itemName = specific_selector.name();
-                if (spix::qt::GetObjectName(item) == QString::fromStdString(itemName)) {
-                    return item;
-                }
-                return nullptr;
-            } else if constexpr (std::is_same_v<SelectorType, spix::path::PropertySelector>) {
-                if (!item) {
-                    return nullptr;
-                }
-                const auto& propertyName = specific_selector.name();
-                QVariant propertyValue = item->property(propertyName.c_str());
-                
-                if (!propertyValue.isValid()) {
-                    return nullptr;
-                }
-                
-                return propertyValue.value<QObject*>();
-            } else {
-                return nullptr; // Unknown selector type
-            }
-        },
-        selector);
-}
-
-/**
- * Performs a DFS to find the first matching item in the UI tree.
- *
- * @param pathComponents Vector of path components to match
- * @param currentNode Starting node for the search
- * @param matchedCount Number of path components already matched in the ancestor chain
- * @return The first matching QQuickItem or nullptr if none found
- */
-QQuickItem* FindMatchingItem(
-    const std::vector<spix::path::Component>& pathComponents, QObject* currentNode, size_t matchedCount)
-{
-    if (!currentNode) {
-        return nullptr;
-    }
-    if (matchedCount >= pathComponents.size()) {
-        return nullptr;
-    }
-
-    // If we have a potential match, check if this node matches the next component
-    const auto& component = pathComponents[matchedCount];
-    const auto& selector = component.selector();
-
-    // Check if this node matches the current selector
-    QObject* matchedObject = MatchesSelector(currentNode, selector);
-    if (matchedObject) {
-        // Increment matched count as we found a match
-        matchedCount++;
-        
-        // If we've matched all components, return this item if it's a QQuickItem
-        if (matchedCount == pathComponents.size()) {
-            return qobject_cast<QQuickItem*>(matchedObject);
-        }
-        
-        const auto& nextComponent = pathComponents[matchedCount];
-        // Continue searching in the matched object's hierarchy
-        // if it is different from the current node or if the next component is a property selector,
-        // as a property selector might reference a property of the current node
-        if (matchedObject != currentNode || std::holds_alternative<spix::path::PropertySelector>(nextComponent.selector())) {
-            return FindMatchingItem(pathComponents, matchedObject, matchedCount);
-        }
-    }
-
-    // Continue DFS through children
-    QQuickItem* result = nullptr;
-
-    // Use ForEachChild to iterate through all children, with special handling for repeaters
-    spix::qt::ForEachChild(currentNode, [&](QObject* child) -> bool {
-        if ((result = FindMatchingItem(pathComponents, child, matchedCount))) {
-            return false; // Stop iteration if we found a match
-        }
-        return true; // Continue iteration
-    });
-
-    return result; // Will return nullptr if no match was found
-}
-
-QQuickItem* GetQQuickItemAtPath(const spix::ItemPath& path)
-{
-    if (path.length() == 0) {
-        return nullptr;
-    }
-
-    // First find the window
-    const auto& windowComponent = path.rootComponent();
-    const auto& windowSelector = windowComponent.selector();
-
-    if (!std::holds_alternative<spix::path::NameSelector>(windowSelector)) {
-        return nullptr;
-    }
-
-    const auto& windowName = std::get<spix::path::NameSelector>(windowSelector).name();
-    QQuickWindow* window = GetQQuickWindowWithName(windowName);
-
-    if (!window) {
-        return nullptr;
-    }
-
-    // If path only has window component, return window's contentItem
-    if (path.length() == 1) {
-        return window->contentItem();
-    }
-
-    // Start DFS from window's contentItem to find the item
-    // Skip the window component (index 0) and start matching from the first child component
-    auto components = path.components();
-    return FindMatchingItem(components, window->contentItem(), 1);
-}
-
-} // namespace
 
 namespace spix {
 
 std::unique_ptr<Item> QtScene::itemAtPath(const ItemPath& path)
 {
-    if (path.length() == 0) {
-        return {};
-    }
-
-    // Find window first
-    QQuickWindow* window
-        = GetQQuickWindowWithName(std::get<path::NameSelector>(path.rootComponent().selector()).name());
-
-    if (!window) {
-        return {};
-    }
-
-    // If only window in path, return window
-    if (path.length() == 1) {
-        return std::make_unique<QtItem>(window);
-    }
-
-    // Find item within window via DFS
-    auto components = path.components();
-    auto item = FindMatchingItem(components, window->contentItem(), 1);
+    // Find item within window
+    auto item = qt::GetQQuickItemAtPath(path);
 
     if (!item) {
         return {};
@@ -196,7 +40,7 @@ Events& QtScene::events()
 
 void QtScene::takeScreenshot(const ItemPath& targetItem, const std::string& filePath)
 {
-    auto item = GetQQuickItemAtPath(targetItem);
+    auto item = qt::GetQQuickItemAtPath(targetItem);
     if (!item) {
         return;
     }
@@ -218,7 +62,7 @@ void QtScene::takeScreenshot(const ItemPath& targetItem, const std::string& file
 
 std::string QtScene::takeScreenshotAsBase64(const ItemPath& targetItem)
 {
-    auto item = GetQQuickItemAtPath(targetItem);
+    auto item = qt::GetQQuickItemAtPath(targetItem);
     if (!item) {
         return "";
     }
